refactor CASR to a more general thing
